Refactoring 2
