+++ examples/test_ecdsa.ino
@@ -1,7 +1,7 @@
 /* Copyright 2014, Kenneth MacKay. Licensed under the BSD 2-clause license. */
 
 #include "uECC.h"
-#include "hmac.h"
+#include "sha256.h"
 #include "keys.h"
 
 #include <stdio.h>
